@@ -9,6 +9,7 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "@hapi/boom": "^9.1.0",
     "@types/node": "^12.7.5",
     "bcrypt": "^3.0.6",
     "body-parser": "^1.19.0",
@@ -25,6 +26,17 @@
     "tsconfig": "^7.0.0",
     "tsconfig-paths": "^3.9.0",
     "typeorm": "^0.2.20",
-    "typescript": "^3.6.2"
+    "typescript": "^4.0.3"
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^4.2.0",
+    "@typescript-eslint/parser": "^4.2.0",
+    "eslint": "^7.10.0",
+    "eslint-config-airbnb-base": "^14.2.0",
+    "eslint-config-prettier": "^6.12.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "eslint-plugin-unused-imports": "^0.1.3",
+    "prettier": "^2.1.2",
+    "prettier-eslint": "^11.0.0"
   }
 }
