@@ -1,11 +1,11 @@
 import { browser, by, element } from 'protractor';
 
 export class AppPage {
-  navigateTo() {
-    return browser.get('/') as Promise<any>;
-  }
+    navigateTo() {
+        return browser.get('/') as Promise<any>;
+    }
 
-  getTitleText() {
-    return element(by.css('app-root h1')).getText() as Promise<string>;
-  }
+    getTitleText() {
+        return element(by.css('app-root h1')).getText() as Promise<string>;
+    }
 }
