@@ -1,15 +1,8 @@
 import { Component } from '@angular/core';
 
-import { UserInterface } from '@interfaces/user.interface'
-
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss'],
 })
-export class AppComponent {
-
-  ngOnInit() {
-  }
-
-}
+export class AppComponent {}
