@@ -1,112 +1,165 @@
-import { Component, OnInit, Input, Output, EventEmitter, ViewChild, SimpleChanges } from '@angular/core';
-import { Observable } from 'rxjs'
-
-import { Form } from '@models/forms/form'
+import {
+    Component,
+    EventEmitter,
+    Input,
+    OnInit,
+    Output,
+    ViewChild,
+} from '@angular/core';
+import { Form, FormFieldUpdated } from '@models/forms/form';
+import { Observable } from 'rxjs';
 
 @Component({
-  selector: 'app-form',
-  templateUrl: './form.component.html',
-  styleUrls: ['./form.component.scss']
+    selector: 'app-form',
+    templateUrl: './form.component.html',
+    styleUrls: ['./form.component.scss'],
 })
 export class FormComponent implements OnInit {
+    @ViewChild('submitButton', { static: false }) submitButton;
 
-	@ViewChild('submitButton', { static: false }) submitButton
+    @Input() form: Form;
 
-	@Input() form: Form
-	@Input() showSubmit: boolean = true // Whether or not to show the submit button. Generally true, save for places like when placed in a modal.
-	@Input() forceSubmit: Observable<any> // If provided, can be used to forcefully submit the form, regardless of user intention
-	@Output() onSubmit: EventEmitter<{}> = new EventEmitter();
-	@Output() onFieldChange: EventEmitter<{ name: string, value: any }> = new EventEmitter();
-	internalStore = {};
+    @Input() showSubmit = true; // Whether or not to show the submit button. Generally true, save for places like when placed in a modal.
 
-	pageNumber: number = 0 // which page number of the form are we currently executing?
+    @Input() forceSubmit: Observable<boolean>; // If provided, can be used to forcefully submit the form, regardless of user intention
 
-	defaultWYSIWYGModules = {
-		formula: true,
-		toolbar: [
-			['bold', 'italic', 'underline', 'strike'],        // toggled buttons
-			['blockquote', 'code-block'],
+    @Output() onSubmit: EventEmitter<any> = new EventEmitter();
 
-			[{ 'list': 'ordered'}, { 'list': 'bullet' }],
-			[{ 'script': 'sub'}, { 'script': 'super' }],      // superscript/subscript
-			[{ 'indent': '-1'}, { 'indent': '+1' }],          // outdent/indent
+    @Output() onFieldChange: EventEmitter<
+        FormFieldUpdated
+    > = new EventEmitter();
 
-			[{ 'size': ['small', false, 'large', 'huge'] }],  // custom dropdown
+    internalStore = {};
 
-			[{ 'color': [] }, { 'background': [] }],          // dropdown with defaults from theme
+    Array = Array;
 
-			['formula', 'link', 'image', 'video']                         // link and image, video
-		]
-	};
+    rowLayouts = {};
 
-	constructor(){}
+    pageNumber = 0; // which page number of the form are we currently executing?
 
-	ngOnInit(){
-		if(this.forceSubmit){
-			this.forceSubmit.subscribe({
-				next: () => {
-					setTimeout(_ => this.submitButton.nativeElement.click())
-				}
-			})
-		}
+    defaultWYSIWYGModules = {
+        formula: true,
+        toolbar: [
+            ['bold', 'italic', 'underline', 'strike'], // toggled buttons
+            ['blockquote', 'code-block'],
 
-		this.loadDefaultFormData()
-	}
+            [{ list: 'ordered' }, { list: 'bullet' }],
+            [{ script: 'sub' }, { script: 'super' }], // superscript/subscript
+            [{ indent: '-1' }, { indent: '+1' }], // outdent/indent
 
-	ngOnChanges(changes: SimpleChanges){
-		this.loadDefaultFormData()
-	}
+            [{ size: ['small', false, 'large', 'huge'] }], // custom dropdown
 
-	loadDefaultFormData(){
-		if(!this.form) return
+            [{ color: [] }, { background: [] }], // dropdown with defaults from theme
 
-		this.form.inputs.forEach((input) => {
-			this.internalStore[input.name] = input.value
-		})
-	}
+            ['formula', 'link', 'image', 'video'], // link and image, video
+        ],
+    };
 
-	// When an editor is created, we pass back the value we wish to save into it in addition to the editor
-	// we then can format the content and then set the contents
-	// Why not do this beforehand? It seems they don't support it very well.
-	// This approach is documented here: https://github.com/KillerCodeMonkey/ngx-quill/issues/77
-	wysiwygCreated(value, editor) {
-		const contents = editor.clipboard.convert(value);
-		editor.setContents(contents)
-	}
+    ngOnInit(): void {
+        if (this.forceSubmit) {
+            this.forceSubmit.subscribe({
+                next: () => {
+                    setTimeout(() => this.submitButton.nativeElement.click());
+                },
+            });
+        }
 
-	nextPage(){
-		if(this.form.pages.length > (this.pageNumber + 1)) this.pageNumber ++
-		else this.submit()
-	}
+        this.loadDefaultFormData();
+    }
 
-	prevPage(){
-		if(this.form.pages.length && this.pageNumber > 0) this.pageNumber --
-	}
+    ngOnChanges(): void {
+        this.loadDefaultFormData();
+    }
 
-	fieldUpdatedWysiwyg(name, data) {
-		const value = (data.target ? data.target.value : data).html
-		this.internalStore[name] = value
+    loadDefaultFormData(): void {
+        if (!this.form) return;
 
+        this.recomputeLayout();
 
-		this.onFieldChange.emit({
-			name,
-			value
-		})		
-	}
+        this.form.inputs.forEach((input) => {
+            this.internalStore[input.name] = input.value;
+        });
+    }
 
-	fieldUpdated(name, data){
-		const value = (data.target ? data.target.value : data)
-		this.internalStore[name] = value
+    // Generates an array where each index is a row, and the value of the index is how many columns that row has
+    recomputeLayout(): void {
+        this.rowLayouts = {};
 
+        // Each group will have its own layouts per group
+        this.form.inputGroups.forEach((group) => {
+            // Find all inputs on this page
+            const inputs = this.form.inputs.filter(
+                (input) =>
+                    input.group === group.name ||
+                    (group.name === '' && !input.group)
+            );
 
-		this.onFieldChange.emit({
-			name,
-			value
-		})
-	}
+            this.rowLayouts[group.name] = [1]; // By default, we have a 1x1 grid, and will fill it out more as needed
 
-	submit(){
-		this.onSubmit.emit(this.internalStore)
-	}
+            inputs.forEach((input) => {
+                // If there isn't a row or col, we assign 0 to the row and col
+                // This will put it into the first grid position
+                // eslint-disable-next-line no-param-reassign
+                input.row = input.row || 0;
+                // eslint-disable-next-line no-param-reassign
+                input.col = input.col || 0;
 
+                // Now we insert the row if needed
+                // This is done by adding an array of (needed length - current length) filled with 1 to the end of the array.
+                if (input.row + 1 > this.rowLayouts[group.name].length)
+                    this.rowLayouts[group.name] = [
+                        ...this.rowLayouts[group.name],
+                        ...Array(
+                            input.row + 1 - this.rowLayouts[group.name].length
+                        ).fill(1),
+                    ];
+
+                // And now we increment the column if needed
+                if (input.col + 1 > this.rowLayouts[group.name][input.row])
+                    this.rowLayouts[group.name][input.row] = input.col + 1;
+            });
+        });
+    }
+
+    // When an editor is created, we pass back the value we wish to save into it in addition to the editor
+    // we then can format the content and then set the contents
+    // Why not do this beforehand? It seems they don't support it very well.
+    // This approach is documented here: https://github.com/KillerCodeMonkey/ngx-quill/issues/77
+    wysiwygCreated(value, editor): void {
+        const contents = editor.clipboard.convert(value);
+        editor.setContents(contents);
+    }
+
+    nextPage(): void {
+        if (this.form.pages.length > this.pageNumber + 1) this.pageNumber += 1;
+        else this.submit();
+    }
+
+    prevPage(): void {
+        if (this.form.pages.length && this.pageNumber > 0) this.pageNumber -= 1;
+    }
+
+    fieldUpdatedWysiwyg(name, data): void {
+        const value = (data.target ? data.target.value : data).html;
+        this.internalStore[name] = value;
+
+        this.onFieldChange.emit({
+            name,
+            value,
+        });
+    }
+
+    fieldUpdated(name, data): void {
+        const value = data.target ? data.target.value : data;
+        this.internalStore[name] = value;
+
+        this.onFieldChange.emit({
+            name,
+            value,
+        });
+    }
+
+    submit(): void {
+        this.onSubmit.emit(this.internalStore);
+    }
 }
