@@ -1,27 +1,22 @@
-import { Component, OnInit } from '@angular/core'
-
-import { UserService } from '@services/user.service'
-import { User } from '@models/user'
+import { Component, OnInit } from '@angular/core';
+import { User } from '@models/user';
+import { UserService } from '@services/user.service';
 
 @Component({
-  selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.scss']
+    selector: 'app-header',
+    templateUrl: './header.component.html',
+    styleUrls: ['./header.component.scss'],
 })
 export class HeaderComponent implements OnInit {
+    user: User = null;
 
-	user: User = null
+    constructor(private userService: UserService) {}
 
-	constructor(
-		private _userService: UserService,
-	) { }
-
-	async ngOnInit() {
-		this._userService.getCurrentUser().subscribe({
-			next: (currentUser) => {
-				this.user = currentUser
-			}
-		})
-	}
-
+    async ngOnInit(): Promise<void> {
+        this.userService.getCurrentUser().subscribe({
+            next: (currentUser) => {
+                this.user = currentUser;
+            },
+        });
+    }
 }
