@@ -1,114 +1,123 @@
-import { Component, OnInit, Input, Output, EventEmitter, SimpleChanges } from '@angular/core';
-import { Subject } from 'rxjs';
-import { ToastrService } from 'ngx-toastr';
-
-import { MeetingTimeService } from '@services/meetingTime.service';
-
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { MeetingType } from '@enums/meetingType.enum';
-import { MeetingTime } from '@models/meetingTime'
-import { Form } from '@models/forms/form'
+import { Form } from '@models/forms/form';
+import { MeetingTime } from '@models/meetingTime';
+import { MeetingTimeService } from '@services/meetingTime.service';
+import { ToastrService } from 'ngx-toastr';
+import { Subject } from 'rxjs';
 
 @Component({
-  selector: 'app-meeting-time-edit',
-  templateUrl: './meeting-time-edit.component.html',
-  styleUrls: ['./meeting-time-edit.component.scss']
+    selector: 'app-meeting-time-edit',
+    templateUrl: './meeting-time-edit.component.html',
+    styleUrls: ['./meeting-time-edit.component.scss'],
 })
 export class MeetingTimeEditComponent implements OnInit {
+    @Input() isVisible: boolean;
 
-	@Input() isVisible: boolean
-	@Input() meetingTime: MeetingTime
-	@Output() onClose: EventEmitter<{}> = new EventEmitter();
-	@Output() onMeetingTimeEdited: EventEmitter<MeetingTime> = new EventEmitter();
-	forceClose: Subject<any> = new Subject<any>()
-	form: Form
+    @Input() meetingTime: MeetingTime;
 
-	constructor(
-		private _meetingTimeService: MeetingTimeService,
-		private toastr: ToastrService,
-	) {}
+    @Output() onClose: EventEmitter<boolean> = new EventEmitter();
 
-	ngOnInit() {
-		this.generateForm()
-	}
+    @Output() onMeetingTimeEdited: EventEmitter<
+        MeetingTime
+    > = new EventEmitter();
 
-	ngOnChanges(changes: SimpleChanges){
-		this.generateForm()
-	}
+    forceClose: Subject<void> = new Subject<void>();
 
-	generateForm(){
-		this.form = new Form();
-		this.form.inputs = [{
-			type: 'select',
-			name: 'type',
-			options: [
-				{ value: MeetingType.RECITATION, label: 'Recitation' },
-				{ value: MeetingType.LECTURE, label: 'Lecture' },
-				{ value: MeetingType.OFFICE_HOUR, label: 'Office Hour' },
-				{ value: MeetingType.STUDY_GROUP, label: 'Study Group' },
-			],
-			value: this.meetingTime.type,
-			label: 'Meeting Type',
-		}, {
-			type: 'select',
-			name: 'weekday',
-			options: [
-				{ value: 'monday', label: 'Monday' },
-				{ value: 'tuesday', label: 'Tuesday' },
-				{ value: 'wednesday', label: 'Wednesday' },
-				{ value: 'thursday', label: 'Thursday' },
-				{ value: 'friday', label: 'Friday' },
-				{ value: 'saturday', label: 'Saturday' },
-				{ value: 'sunday', label: 'Sunday' },
-			],
-			value: this.meetingTime.weekday,
-			label: 'Weekday',
-		}, {
-			type: 'user',
-			name: 'leader',
-			value: this.meetingTime.leader,
-			label: 'Leader',
-		}, {
-			type: 'time',
-			name: 'startTime',
-			value: this.meetingTime.startTime,
-			label: 'Start Time',
-		}, {
-			type: 'time',
-			name: 'endTime',
-			value: this.meetingTime.endTime,
-			label: 'End Time',
-		}, {
-			type: 'select',
-			name: 'frequency',
-			options: [
-				{ value: 1, label: 'Once a week' }
-			],
-			value: this.meetingTime.frequency,
-			label: 'Frequency',
-		}]
-	}
+    form: Form;
 
-	handleClose(){
-		this.onClose.emit(null)
-	}
+    constructor(
+        private meetingTimeService: MeetingTimeService,
+        private toastr: ToastrService
+    ) {}
 
-	async formSubmitted(data){
-		// first we update the data in the model
-		Object.keys(data).forEach((key) => {
-			this.meetingTime[key] = data[key]
-		})
+    ngOnInit(): void {
+        this.generateForm();
+    }
 
-		// and now we submit it to the API.
-		try{
-			const result = await this._meetingTimeService.createMeetingTime(this.meetingTime)
-			this.onMeetingTimeEdited.emit(result)
+    ngOnChanges(): void {
+        this.generateForm();
+    }
 
-			this.toastr.success('Successfully created meeting time')
-			this.forceClose.next()
-		}catch(err){
-			console.error(err)
-			this.toastr.error('Failed to create meeting time')
-		}
-	}
+    generateForm(): void {
+        this.form = new Form();
+        this.form.inputs = [
+            {
+                type: 'select',
+                name: 'type',
+                options: [
+                    { value: MeetingType.RECITATION, label: 'Recitation' },
+                    { value: MeetingType.LECTURE, label: 'Lecture' },
+                    { value: MeetingType.OFFICE_HOUR, label: 'Office Hour' },
+                    { value: MeetingType.STUDY_GROUP, label: 'Study Group' },
+                ],
+                value: this.meetingTime.type,
+                label: 'Meeting Type',
+            },
+            {
+                type: 'select',
+                name: 'weekday',
+                options: [
+                    { value: 'monday', label: 'Monday' },
+                    { value: 'tuesday', label: 'Tuesday' },
+                    { value: 'wednesday', label: 'Wednesday' },
+                    { value: 'thursday', label: 'Thursday' },
+                    { value: 'friday', label: 'Friday' },
+                    { value: 'saturday', label: 'Saturday' },
+                    { value: 'sunday', label: 'Sunday' },
+                ],
+                value: this.meetingTime.weekday,
+                label: 'Weekday',
+            },
+            {
+                type: 'user',
+                name: 'leader',
+                value: this.meetingTime.leader,
+                label: 'Leader',
+            },
+            {
+                type: 'time',
+                name: 'startTime',
+                value: this.meetingTime.startTime,
+                label: 'Start Time',
+            },
+            {
+                type: 'time',
+                name: 'endTime',
+                value: this.meetingTime.endTime,
+                label: 'End Time',
+            },
+            {
+                type: 'select',
+                name: 'frequency',
+                options: [{ value: 1, label: 'Once a week' }],
+                value: this.meetingTime.frequency,
+                label: 'Frequency',
+            },
+        ];
+    }
 
+    handleClose(): void {
+        this.onClose.emit(null);
+    }
+
+    async formSubmitted(data): Promise<void> {
+        // first we update the data in the model
+        Object.keys(data).forEach((key) => {
+            this.meetingTime[key] = data[key];
+        });
+
+        // and now we submit it to the API.
+        try {
+            const result = await this.meetingTimeService.createMeetingTime(
+                this.meetingTime
+            );
+            this.onMeetingTimeEdited.emit(result);
+
+            this.toastr.success('Successfully created meeting time');
+            this.forceClose.next();
+        } catch (err) {
+            this.toastr.error('Failed to create meeting time');
+        }
+    }
 }
