@@ -1,70 +1,92 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { BehaviorSubject } from 'rxjs';
-
+import { Course } from '@models/course';
+import { Section } from '@models/section';
 import { CourseService } from '@services/course.service';
 import { SectionService } from '@services/section.service';
 
-import { Course } from '@models/course'
-import { Section } from '@models/section'
-
 @Component({
-  selector: 'app-view-sections',
-  templateUrl: './view-sections.component.html',
-  styleUrls: ['./view-sections.component.scss']
+    selector: 'app-view-sections',
+    templateUrl: './view-sections.component.html',
+    styleUrls: ['./view-sections.component.scss'],
 })
 export class ViewSectionsComponent implements OnInit {
+    course: Course;
 
-  	course: Course
-  	sections: Section[]
-  	isLoading: boolean = true
+    sections: Section[];
 
-  	selectedEditSection: Section = null
+    isLoading = true;
 
-  	isEditSectionModalOpen: boolean = false
+    selectedEditSection: Section = null;
 
-	constructor(
-		private _courseService: CourseService,
-		private _sectionService: SectionService,
-		private route: ActivatedRoute,
-	) { }
+    selectedDeleteSection: Section = null;
 
-	ngOnInit() {
-		this.route.params.subscribe(async (params) => {
-			if(params['courseID']) {
-				this.course = await this._courseService.getCourse(params['courseID'])
-				this.sections = await this._sectionService.getCourseSections(this.course)
-				this.isLoading = false
-			}
-		});
-	}
+    isEditSectionModalOpen = false;
 
-	handleOpenNewSectionModal() {
-		this.isEditSectionModalOpen = true
+    isDeleteSectionModalOpen = false;
 
-		this.selectedEditSection = new Section()
-		this.selectedEditSection.course = this.course;
-	}
-	
-	handleCloseEditSectionModal() {
-		this.isEditSectionModalOpen = false
+    constructor(
+        private courseService: CourseService,
+        private sectionService: SectionService,
+        private route: ActivatedRoute
+    ) {}
 
-		// And now we add the section if needed
-		// We perform a search for if there is a section with that id already
-		const foundSection = this.sections.find((section) => {
-			if(section.id == this.selectedEditSection.id) return true
-		})
+    ngOnInit(): void {
+        this.route.params.subscribe(async (params) => {
+            if (params.courseID) {
+                this.course = await this.courseService.getCourse(
+                    params.courseID
+                );
+                this.sections = await this.sectionService.getCourseSections(
+                    this.course
+                );
+                this.isLoading = false;
+            }
+        });
+    }
 
-		// if the section was found, we already have it in our array, and the data would be updated via the component
-		// if it wasn't found, we insert it new.
-		if(!foundSection) this.sections.push(this.selectedEditSection)
+    handleOpenNewSectionModal(): void {
+        this.isEditSectionModalOpen = true;
 
-		this.selectedEditSection = null;
-	}
+        this.selectedEditSection = new Section();
+        this.selectedEditSection.course = this.course;
+    }
 
-	handleOpenEditSectionModal(section: Section) {
-		this.isEditSectionModalOpen = true
-		this.selectedEditSection = section
-	}
+    handleCloseEditSectionModal(): void {
+        this.isEditSectionModalOpen = false;
 
+        // And now we add the section if needed
+        // We perform a search for if there is a section with that id already
+        const foundSection = this.sections.find((section) => {
+            return section.id === this.selectedEditSection.id;
+        });
+
+        // if the section was found, we already have it in our array, and the data would be updated via the component
+        // if it wasn't found, we insert it new.
+        if (!foundSection) this.sections.push(this.selectedEditSection);
+
+        this.selectedEditSection = null;
+    }
+
+    handleOpenEditSectionModal(section: Section): void {
+        this.isEditSectionModalOpen = true;
+        this.selectedEditSection = section;
+    }
+
+    handleOpenDeleteSectionModal(section: Section): void {
+        // let result = this.sectionService.deleteSection(section.id);
+        this.isDeleteSectionModalOpen = true;
+        this.selectedDeleteSection = section;
+    }
+
+    handleCloseDeleteSectionModal($event): void {
+        this.isDeleteSectionModalOpen = false;
+
+        if ($event) {
+            this.sections.splice(
+                this.sections.indexOf(this.selectedDeleteSection),
+                1
+            );
+        }
+    }
 }
