@@ -1,60 +1,73 @@
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 
-import { BehaviorSubject, Observable } from 'rxjs'
+import { environment } from '@environment';
 
-import { environment } from '@environment'
-
-import { Course } from '@models/course'
-import { Section } from '@models/section'
+import { Course } from '@models/course';
+import { Section } from '@models/section';
 
 @Injectable()
 export class SectionService {
+    constructor(private http: HttpClient) {}
 
-	constructor(
-		private http: HttpClient
-	){}
+    public async upsertSection(section: Section): Promise<Section> {
+        const sectionID = section.id;
+        const url = sectionID
+            ? `${environment.apiURL}/section/${sectionID}`
+            : `${environment.apiURL}/section`;
+        let action;
 
-	public getHeaders(){
-		return new HttpHeaders({
-			'Content-Type':  'application/json',
-			'Authorization': 'Bearer ' + localStorage.getItem('jwt')
-		})
-	}
+        if (sectionID) action = this.http.put(url, section);
+        else action = this.http.post(url, section);
 
-	public async upsertSection(section: Section) : Promise<Section>{
-		const sectionID = section.id
-		const url = (sectionID) ? `${environment.apiURL}/section/${sectionID}` : `${environment.apiURL}/section`
-		let action
+        return new Promise((resolve, reject) => {
+            action.subscribe(
+                (result: { data: Section }) => {
+                    if (result) resolve(result.data);
+                    else reject(new Error('No result returned'));
+                },
+                (err) => {
+                    reject(err);
+                }
+            );
+        });
+    }
 
-		if(sectionID) action = this.http.put(url, section, { headers: this.getHeaders() })
-		else action = this.http.post(url, section, { headers: this.getHeaders() })
+    public async getCourseSections(
+        course: Course | string
+    ): Promise<Section[]> {
+        // if course is an object, we will grab its id
+        // otherwise, we assume course is a string representing the id
+        const courseID = course instanceof String ? course : course.id;
 
-		return new Promise((resolve, reject) => {
-			action.subscribe((result: { data: Section }) => {
-				if(result) resolve(result.data)
-				else reject(new Error("No result returned"))
-			}, (err) => {
-				reject(err)
-			})
-		})
-	}
+        const url = `${environment.apiURL}/course/${courseID}/sections`;
 
-	public async getCourseSections(course: Course | String) : Promise<Section[]>{
-		// if course is an object, we will grab its id
-		// otherwise, we assume course is a string representing the id
-		const courseID = (course instanceof String) ? course : course.id
+        return new Promise((resolve, reject) => {
+            this.http.get(url).subscribe(
+                (result: { data: Section[] }) => {
+                    if (result) resolve(result.data);
+                    else reject(new Error('No result returned'));
+                },
+                (err) => {
+                    reject(err);
+                }
+            );
+        });
+    }
 
-		const url = `${environment.apiURL}/course/${courseID}/sections`
+    public async deleteSection(sectionID: string) {
+        const url = `${environment.apiURL}/section/${sectionID}`;
 
-		return new Promise((resolve, reject) => {
-			this.http.get(url, { headers: this.getHeaders() }).subscribe((result: { data: Section[] }) => {
-				if(result) resolve(result.data)
-				else reject(new Error("No result returned"))
-			}, (err) => {
-				reject(err)
-			})
-		})
-	}
-
+        return new Promise((resolve, reject) => {
+            this.http.delete(url).subscribe(
+                (result: { data: Section }) => {
+                    if (result) resolve(result.data);
+                    else reject(new Error('No result returned'));
+                },
+                (err) => {
+                    reject(err);
+                }
+            );
+        });
+    }
 }
