@@ -3,9 +3,9 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  authMechanism: 'manual',
-  apiURL: 'http://localhost:3000'
+    production: false,
+    authMechanism: 'manual',
+    apiURL: 'http://localhost:3000',
 };
 
 /*
