@@ -1,5 +1,6 @@
 export enum MeetingType {
-
-	RECITATION, LECTURE, OFFICE_HOUR, STUDY_GROUP
-
+    RECITATION,
+    LECTURE,
+    OFFICE_HOUR,
+    STUDY_GROUP,
 }
