@@ -1,20 +1,18 @@
-import { CourseInterface } from '@interfaces/course.interface'
-import { UserInterface } from '@interfaces/user.interface'
-import { MeetableInterface } from '@interfaces/meetable.interface'
+import { CourseInterface } from '@interfaces/course.interface';
+import { UserInterface } from '@interfaces/user.interface';
+import { MeetableInterface } from '@interfaces/meetable.interface';
 
 export interface SectionInterface extends MeetableInterface {
+    id: string; // the document id
 
-	id: string;			// the document id
+    index: string; // The index code of this section
+    sectionNumber: string; // The string numerically identifying this section.
 
-	index: string;			// The index code of this section
-	sectionNumber: string;	// The string numerically identifying this section.
+    // students: UserInterface[]; 		// All of the students registered to this section
 
-	// students: UserInterface[]; 		// All of the students registered to this section
+    course: CourseInterface; // the course that the section belongs to
 
-	course: CourseInterface; 	// the course that the section belongs to 
-
-	ta?: UserInterface; 	// the TA for the section
-	professor?: UserInterface;		// the professor for the section
-	students?: UserInterface[];		// the students for the section
-
+    ta?: UserInterface; // the TA for the section
+    instructor?: UserInterface; // the instructor for the section
+    students?: UserInterface[]; // the students for the section
 }
