@@ -1,36 +1,47 @@
-import { Entity, PrimaryGeneratedColumn, Column, BaseEntity, OneToMany, JoinTable } from 'typeorm'
+import {
+    Entity,
+    PrimaryGeneratedColumn,
+    Column,
+    BaseEntity,
+    OneToMany,
+    JoinTable,
+} from 'typeorm';
 
-import { CourseInterface } from '@interfaces/course.interface'
-import { User } from '@models/user'
-import { Section } from '@models/section'
+import { CourseInterface } from '@interfaces/course.interface';
+
+import { Section } from '@models/section';
 
 @Entity()
 export class Course extends BaseEntity implements CourseInterface {
+    @PrimaryGeneratedColumn('uuid')
+    public id: string;
 
-	@PrimaryGeneratedColumn("uuid")
-	public id: string;
+    @Column()
+    public name: string;
 
-	@Column()
-	public name: string;
+    @Column()
+    public department: string;
 
-	@Column()
-	public department: string
+    @Column()
+    public courseCode: string;
 
-	@Column()
-	public courseCode: string
+    @OneToMany((type) => Section, (section) => section.course, {
+        eager: true,
+        cascade: true,
+    })
+    @JoinTable()
+    public sections?: Section[];
 
-	@OneToMany(type => Section, section => section.course, { eager: true, cascade: true })
-	@JoinTable()
-	public sections?: Section[]
-
-	constructor(args: {
-		id?: string,
-		name?: string,
-		department?: string,
-		courseCode?: string,
-		sections?: Section[],
-	} = {}){
-		super()
-		Object.assign(this, args)
-	}
+    constructor(
+        args: {
+            id?: string;
+            name?: string;
+            department?: string;
+            courseCode?: string;
+            sections?: Section[];
+        } = {}
+    ) {
+        super();
+        Object.assign(this, args);
+    }
 }
