@@ -1,46 +1,58 @@
-import { ChildEntity, PrimaryGeneratedColumn, Column, BaseEntity, ManyToMany, OneToOne, JoinTable, JoinColumn, OneToMany, ManyToOne } from 'typeorm'
+import {
+    ChildEntity,
+    PrimaryGeneratedColumn,
+    Column,
+    ManyToMany,
+    JoinTable,
+    JoinColumn,
+    ManyToOne,
+} from 'typeorm';
 
-import { SectionInterface } from '@interfaces/section.interface'
-import { User } from '@models/user'
-import { Course } from '@models/course'
-import { MeetingTime } from '@models/meetingTime'
-import { Meetable } from '@models/meetable'
+import { SectionInterface } from '@interfaces/section.interface';
+import { User } from '@models/user';
+import { Course } from '@models/course';
+
+import { Meetable } from '@models/meetable';
 
 @ChildEntity()
 export class Section extends Meetable implements SectionInterface {
-	
-	@Column()
-	public index: string
+    @PrimaryGeneratedColumn('uuid')
+    public id: string;
 
-	@Column()
-	public sectionNumber: string
+    @Column()
+    public index: string;
 
-	@ManyToOne(type => Course, course => course.sections)
-	@JoinColumn()
-	public course: Course
+    @Column()
+    public sectionNumber: string;
 
-	@ManyToOne(type => User, { eager: true, cascade: true })
+    @ManyToOne((type) => Course, (course) => course.sections)
     @JoinColumn()
-	public ta?: User
-	
-	@ManyToOne(type => User, { eager: true, cascade: true })
-    @JoinColumn()
-	public professor?: User
+    public course: Course;
 
-	@ManyToMany(type => User)
+    @ManyToOne((type) => User, { eager: true, cascade: true })
+    @JoinColumn()
+    public ta?: User;
+
+    @ManyToOne((type) => User, { eager: true, cascade: true })
+    @JoinColumn()
+    public instructor?: User;
+
+    @ManyToMany((type) => User)
     @JoinTable()
-	public students?: User[]
-	
-	constructor(args: {
-		id?: string,
-		index?: string,
-		sectionNumber?: string,
-		course?: Course,
-		students?: User[],
-		ta?: User,
-		professor?: User
-	} = {}){
-		super()
-		Object.assign(this, args)
-	}
+    public students?: User[];
+
+    constructor(
+        args: {
+            id?: string;
+            index?: string;
+            sectionNumber?: string;
+            course?: Course;
+            students?: User[];
+            ta?: User;
+            instructor?: User;
+        } = {}
+    ) {
+        super();
+        Object.assign(this, args);
+    }
 }
